Add script to update ignorePaths in renovate.json
We want to enable Renovate bot in this repository.

However, this creates a problem. If the shared-ci users (those who
include the templates from this repo in their own repo) have Renovate
enabled as well, they will often get merge conflicts from 'cruft update'
due to having different GH actions versions compared to this repo.

To solve the problem, write a script that finds all the shared-ci
workflows in the repo and updates renovate.json to ignore them all.

The intention is for shared-ci users to use this script to update their
renovate.json files. This will be automated in future commits.

The script will *not* be used in this repo itself. The state that we
want to achieve is:

* this repo gets Renovate updates for the shared CI workflows
* shared-ci users do not get Renovate updates for the same workflows
* they get the updates from from the 'cruft update' workflow, which
  updates their files to match the ones in this repo

Signed-off-by: Adam Cmiel <acmiel@redhat.com>