update-shared-ci: run renovate-ignore script
While updating the shared CI, run the renovate-ignore script in case the
update brings in new workflow files. That way, the new workflows will be
automatically added to renovate.json.

Corner case: if the update causes a merge conflict with the
renovate-ignore script, don't run the script and mention the problem in
the PR description.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>