add task-lint workflow
I've added the workflow from build-definitions, that was in .tekton/tasks/taks-lint.yaml and added the script into the GH workflow

I've set the `yq` version to the v4.47.1 (latest) and didn't use `latest` version, as it is a good practise.
The only change from the original script is, that I've added an additional echo, explaining *why* the check is needed.

Signed-off-by: Filip Mudry <fmudry@redhat.com>